+++ client/src/App.js
@@ -39,7 +39,6 @@ const client = new ApolloClient({
 
 function App() {
   return (
-
     <ApolloProvider client={client}>
       <Router>
         <div>
@@ -56,7 +55,6 @@ function App() {
         </div>
       </Router>
     </ApolloProvider>
-
     <div className="App">
       <Header />
       <Body />
