+++ client/src/utils/auth.js
@@ -1,0 +1,30 @@
+const jwt = require('jsonwebtoken');
+
+const secret = 'mysecret';
+const expiration = '4h';
+
+module.exports = {
+    authMiddleware: function ({ req }) {
+        let token = req.body.token || req.query.token || req.headers.authorization;
+
+        if (req.headers.authorization) {
+            token = token.split(' ').pop().trim();
+        }
+        if (!token) {
+            return req;
+        }
+
+        try {
+            const { data } = jwt.verify(token, secret, { maxAge: expiration });
+            req.user = data;
+        } catch {
+            console.log('Invalid token, check auth.js');
+        }
+        return req;
+    },
+    signToken: function ({ _id, name, email }) {
+        const payload = { _id, name, email };
+
+        return jwt.sign({ data: payload }, secret, { expiresIn: expiration });
+    },
+};
