+++ client/src/utils/mutations.js
@@ -12,33 +12,16 @@ export const LOGIN = gql`
 `;
 
 export const ADD_USER = gql`
-  mutation addUser(
-    $firstName: String!
-    $lastName: String!
-    $email: String!
-    $password: String!
-  ) {
-    addUser(
-      firstName: $firstName
-      lastName: $lastName
-      email: $email
-      password: $password
-    ) {
-      token
-      user {
-        _id
-      }
-    }
-  }
-`;
-
-export const ADD_USER = gql`
 mutation addUser(
   $firstName: String!
   $lastName: String!
   $email: String!
-  $password: String1
-)
+  $password: String!
+){
+  token
+  user {
+    _id
+  }
 `;
 
 export const UPDATE_USER = gql`
